+++ CHANGES.rst
@@ -10,6 +10,8 @@
 - Fix bug where "vector" (shape (n,) not shape (1, n)) arrays would loose all their entries except the
   first if ``with_units=True`` was used. [#563]
 
+- Fix bug when comparing float/int to Quantity in computing inverse. [#582]
+
 0.24.0 (2025-02-04)
 -------------------